[S05P12B101-25] [S05P12B101-80] 방 수정, 삭제 API 구현

+++ backend/src/main/java/com/ssafy/db/entity/Room.java
@@ -26,7 +26,7 @@ import org.checkerframework.common.aliasing.qual.Unique;
 public class Room{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Integer rid = null;
+    Long rid = null;
     
     String name;
     Integer host_id;
